Ref: Created /tables and moved files inside
