+++ app/src/components/market/sections/common_sections/tables/new_value/img/ArrowIcon.tsx
