+++ app/src/components/market/sections/market_pooling/market_pool_liquidity.tsx
@@ -23,7 +23,6 @@ import { BigNumberInputReturn } from '../../../common/form/big_number_input'
 import { ModalTransactionWrapper } from '../../../modal'
 import { CurrenciesWrapper, GenericError, TabsGrid } from '../../common/common_styled'
 import { CurrencySelector } from '../../common/currency_selector'
-import { OutcomeTable } from '../../common/outcome_table'
 import { TokenBalance } from '../../common/token_balance'
 import { TransactionDetailsCard } from '../../common/transaction_details_card'
 import { TransactionDetailsLine } from '../../common/transaction_details_line'
@@ -31,6 +30,7 @@ import { TransactionDetailsRow, ValueStates } from '../../common/transaction_det
 import { SetAllowance } from '../common_sections/allowance/set_allowance'
 import { GridTransactionDetails } from '../common_sections/card_bottom_details/grid_transaction_details'
 import { WarningMessage } from '../common_sections/message_text/warning_message'
+import { OutcomeTable } from '../common_sections/tables/outcome_table'
 
 import { UserPoolData } from './user_pool_data'
